fix foreign key hospitals

+++ database/migrations/2020_04_09_090908_create_hospitals_table.php
@@ -15,7 +15,7 @@ class CreateHospitalsTable extends Migration
     {
         Schema::create('hospitals', function (Blueprint $table) {
             $table->id();
-            $table->integer('area_id');
+            $table->unsignedBigInteger('area_id');
             $table->string('name');
             $table->string('description');
             $table->string('address');
